+++ blueprints/agentic/multi-agent-strands-bedrock/weather/.amazonq/mcp.json
@@ -2,7 +2,7 @@
   "mcpServers": {
     "weather": {
       "command": "uvx",
-      "args": ["--no-cache","--from",".","--directory",".","weather-agent-mcp-server"]
+      "args": ["--from",".","--directory",".","weather-agent-mcp-server"]
     }
   }
 }
