+++ blueprints/agentic/multi-agent-strands-bedrock/weather/agent_a2a_server.py
@@ -1,6 +1,6 @@
 from strands import Agent
 from multiagent.a2a import A2AAgent
-from weather_agent import get_weather_agent
+from agent import get_weather_agent
 
 def weather_a2a_server():
     strands_agent = get_weather_agent()
