+++ blueprints/agentic/multi-agent-strands-bedrock/weather/agent_interactive.py
@@ -1,4 +1,4 @@
-from weather_agent import weather_assistant
+from agent import weather_assistant
 from rich.markdown import Markdown
 from rich.console import Console
